tidy up TextInputWithCurrencySymbol refactor
